データベース構築の大枠を作成
