--- app/models/database.py
@@ -1,6 +1,0 @@
-from sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-
-def init_db(app):
-    db.init_app(app)
